Create Number_PalindromeManipulation.java
Takes an Integer n( 1<=n<1000) as an input and prints it on the screen if it is a palindrome. If not, you add the value  of the number read from   left-to-right and its value read from right-to-left, and prints it on the screen if it is a palindrome. If  not,continues until a palindrome is obtained and printed or prints "Too long" if the addition operations exceeds 25.